Tile collision for horizontal and vertical move.
The algorithm is a bit glitchy but gives an idea of how we might implement
this for our game. The main alrogithm is in MoveSprite in the main game
class (Game1.cs).
Press F5 to build and test the game. Use Left and Right to move, Up to
jump. Use WASD to scroll the map.